backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
requests==2.31.0
websockets==12.0
twilio==8.10.0
supabase==2.3.0
azure-cognitiveservices-speech==1.34.0
openai==1.6.1
aiofiles==23.2.1
python-multipart==0.0.6
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
